refactor: restructure LilyPond pipeline into lilypond/ directory and add CI path filters (#474)
