+++ .github/workflows/ci.yml
@@ -5,6 +5,13 @@ on:
     branches: [main]
   pull_request:
     branches: [main]
+    paths-ignore:
+      - "doc/**"
+      - "*.md"
+      - "lilypond/**"
+      - ".github/workflows/lilypond.yml"
+      - ".github/workflows/e2e.yml"
+      - ".github/workflows/monitor-resources.yml"
   # Daily at 00:00 UTC — catches drift in the test suite on days when no PR
   # triggered the job. GitHub may delay scheduled runs by up to ~15 min during
   # peak load.
