+++ lilypond/test/buildScores.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { parseArgs, buildPattern } from "../../build/buildScores.mjs";
+import { parseArgs, buildPattern } from "../../lilypond/build/buildScores.mjs";
 
 describe("parseArgs", () => {
   it("returns defaults when given no arguments", () => {
@@ -50,14 +50,14 @@ describe("parseArgs", () => {
 
 describe("buildPattern", () => {
   it("returns a wildcard pattern when choir is omitted", () => {
-    expect(buildPattern("src/lilypond/Hugh Keyte/early")).toBe(
-      "src/lilypond/Hugh Keyte/early/Choir*.ly"
+    expect(buildPattern("lilypond/src/Hugh Keyte/early")).toBe(
+      "lilypond/src/Hugh Keyte/early/Choir*.ly"
     );
   });
 
   it("returns a single-choir pattern when choir is provided", () => {
-    expect(buildPattern("src/lilypond/Hugh Keyte/early", "I A")).toBe(
-      "src/lilypond/Hugh Keyte/early/Choir I A.ly"
+    expect(buildPattern("lilypond/src/Hugh Keyte/early", "I A")).toBe(
+      "lilypond/src/Hugh Keyte/early/Choir I A.ly"
     );
   });
 });
