+++ src/ts/lily.ts
@@ -5,7 +5,7 @@ import config from "./config";
 import lyGrammar from "../ohmjs/ly-grammar.ohm-bundle";
 import * as ohm from "ohm-js";
 import { Duration, BarLine, Note, Rest, Component } from "./music-classes";
-import spem from "../lilypond/Hugh Keyte/spem.ly?raw";
+import spem from "../../lilypond/src/Hugh Keyte/spem.ly?raw";
 
 // Make an dictionary of music positions (hemidemisemiquavers/128) to array of notes {choir, part, note}
 export type Dictionary = {
